One way of fixing the message-in-an-action defect.
I sidestepped the problem by anchoring the msg and action patterns
at the beginning of the line, so a message won't match in the
middle of a valid action line just because that pattern happens to
be matched first in parse_line().

The obvious other way to do would have been to swap the order of
the matchings.  This ought to have done the trick too, since what
really makes message logs and action logs differ when all optional
items are out is the presence of the <> markers around the nick.
I decided against that to preserve the assumed most common case,
messages, first.